requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyyaml==6.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
